feat(wave6-11): complete full roadmap — renderer features, build fixes, security, licensing, cloud, chat
Wave 6: Command palette (cmdk), keyboard shortcuts, case name/settings/export
modals, light theme CSS, font size scaling
Wave 7: Fix CEF 145 API breaks, PIMPL destructors, raw string literals,
pe-parse unique_ptr, posix_spawn, macOS environ — clean 4.3MB arm64 build
Wave 8: Wire email analyzer + log analysis IPC handlers
Wave 9: LLVM obfuscation CMake, VMProtect CMake, anti-debug mesh (12 methods),
integrity guard, encrypted pages, rule crypto, hardware fingerprint
Wave 10: License manager (Keychain/DPAPI/Secret Service, 30-day offline grace),
cloud analysis client, encrypted rule sync
Wave 11: 10 ShieldCrypt chat IPC actions, renderer chat store, push events

+++ src/native/analysis/advanced/script_analyzer.cpp
@@ -4,6 +4,8 @@
 #include <regex>
 #include <string>
 
+#include "common/json.h"
+
 namespace shieldtier {
 
 namespace {
@@ -234,7 +236,7 @@ std::vector<Finding> ScriptAnalyzer::detect_base64_payload(const std::string& co
     size_t max_len = 0;
     for (auto it = begin; it != end; ++it) {
         ++count;
-        max_len = std::max(max_len, it->length());
+        max_len = std::max(max_len, static_cast<size_t>(it->length()));
     }
 
     if (count > 0) {
